requirements.txt
transformers
qwen_vl_utils
torch
torchvision
torchaudio
accelerate
streamlit
pillow